added some view models tests. fixed note id bug. need to add to searchBy option to open preview if searching by content?

+++ app/src/main/java/id/slava/nt/simplenotepad/domain/usecase/DeleteNote.kt
@@ -1,5 +1,6 @@
 package id.slava.nt.simplenotepad.domain.usecase
 
+import id.slava.nt.simplenotepad.domain.models.InvalidNoteException
 import id.slava.nt.simplenotepad.domain.models.Note
 import id.slava.nt.simplenotepad.domain.repository.NoteRepository
 
@@ -7,7 +8,14 @@ class DeleteNote(
     private val repository: NoteRepository
 ) {
 
-    suspend operator fun invoke(note: Note) {
-        repository.deleteNote(note)
+    suspend operator fun invoke(note: Note?) {
+        if (note == null){
+
+            throw InvalidNoteException("Note does not exist")
+
+        } else{
+            repository.deleteNote(note)
+        }
+
     }
 }
